add authorization Gate for image upload

+++ app/Http/Controllers/ImageController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Image;
 use Illuminate\Http\Request;
 use Storage;
+use Gate;
 class ImageController extends Controller
 {
     /**
@@ -80,6 +81,11 @@ class ImageController extends Controller
      */
     public function destroy(Image $image)
     {
+
+        if (Gate::denies('delete_image', $image)) {
+            abort(401);
+        }
+
        if (Storage::delete('public/images/' . $image->path) ) {
 
          $image->delete(); //delete image form db
